kms/python-service/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# OCR & PDF
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0

# NLP & Embeddings
sentence-transformers==3.2.1
torch==2.5.1+cpu
torchvision==0.20.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Vector store
chromadb==0.5.18

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
